GetWorkers moved back in internal dir
